langchain/streamlit_LLAMA_notGoodOutput/requirements.txt
streamlit
python-dotenv

PyPDF2

langchain
langchain_community
langchain_openai

faiss-cpu
